Avoid nuxt-link for problems while diffing
