+++ pages/advocates/index.vue
@@ -6,7 +6,7 @@
     </header>
     <main>
       <header>
-        <CanvasHeader
+        <GatesHeader
           id="presentation"
           main-title="Qiskit Advocates"
         >
@@ -39,7 +39,7 @@
               </CompactFeature>
             </div>
           </template>
-        </CanvasHeader>
+        </GatesHeader>
       </header>
       <div class="inner-navigation-scope">
         <InnerNavigation
@@ -52,13 +52,11 @@
         />
         <PageSection id="steps-to-apply">
           <h2>Steps to apply</h2>
-          <p>
-            <ol>
-              <li>Click on the "Apply now" button below.</li>
-              <li>Complete the test attached to the application form.</li>
-              <li>Upload the test results to the application form and hit submit.</li>
-            </ol>
-          </p>
+          <ol>
+            <li>Click on the "Apply now" button below.</li>
+            <li>Complete the test attached to the application form.</li>
+            <li>Upload the test results to the application form and hit submit.</li>
+          </ol>
           <ul class="actions">
             <li>
               <Cta to="https://www.youtube.com/playlist?list=PLOFEBzvs-Vvp2xg9-POLJhQwtVktlYGbY">
@@ -67,10 +65,9 @@
             </li>
           </ul>
         </PageSection>
-        <PageSection id="around-the-world" extra-position="start">
+        <MapSection id="around-the-world">
           <h2>Around the world</h2>
-          <div id="advocates-map" />
-        </PageSection>
+        </MapSection>
         <PageSection id="meet-the-advocates">
           <h2>Meet the advocates</h2>
           <div class="advocate-cards-container">
@@ -95,8 +92,9 @@ import Vue from 'vue'
 import { Component } from 'vue-property-decorator'
 import QiskitOrgMenu from '~/components/menus/QiskitOrgMenu.vue'
 import InnerNavigation from '~/components/menus/InnerNavigation.vue'
-import CanvasHeader from '~/components/headers/CanvasHeader.vue'
+import GatesHeader from '~/components/headers/GatesHeader.vue'
 import PageSection from '~/components/sections/PageSection.vue'
+import MapSection from '~/components/advocates/MapSection.vue'
 import PageFooter from '~/components/footers/PageFooter.vue'
 import Cta from '~/components/ctas/Cta.vue'
 import AdvocateCard from '~/components/cards/AdvocateCard.vue'
@@ -108,8 +106,9 @@ import CompactFeature from '~/components/features/CompactFeature.vue'
   components: {
     QiskitOrgMenu,
     InnerNavigation,
-    CanvasHeader,
+    GatesHeader,
     PageSection,
+    MapSection,
     PageFooter,
     Cta,
     AdvocateCard,
@@ -133,44 +132,6 @@ import CompactFeature from '~/components/features/CompactFeature.vue'
       renderFn: index.vue.render,
       staticRenderFns: index.vue.staticRenderFns
     }
-  },
-
-  mounted() {
-    require('d3')
-    require('topojson')
-    const Datamap = require('datamaps')
-    const map = new Datamap({
-      element: document.getElementById('advocates-map'),
-      responsive: true,
-      projection: 'mercator',
-      geographyConfig: {
-        highlightOnHover: false,
-        popupOnHover: false,
-        borderColor: 'var(--primary-color-lightmost)'
-      },
-      fills: {
-        defaultFill: '#0000',
-        city: 'var(--secondary-color)'
-      },
-      bubblesConfig: {
-        borderWidth: 0,
-        popupTemplate(_, data) {
-          return `<div class="map-tip">${data.name}</div>`
-        },
-        highlightFillColor: 'var(--secondary-color-lightmost)'
-      }
-    })
-    map.bubbles([
-      { name: 'Madrid', latitude: 40.41, longitude: -3.7, fillKey: 'city', radius: 5 },
-      { name: 'New York', latitude: 40.71, longitude: -74, fillKey: 'city', radius: 5 },
-      { name: 'Basel', latitude: 47.56, longitude: 7.5, fillKey: 'city', radius: 5 },
-      { name: 'Whashington D.C.', latitude: 38.89, longitude: -77.01, fillKey: 'city', radius: 5 },
-      { name: 'Surrey', latitude: 51.31, longitude: -0.55, fillKey: 'city', radius: 5 },
-      { name: 'Tokyo', latitude: 35.68, longitude: 139.69, fillKey: 'city', radius: 5 },
-      { name: 'Rome', latitude: 41.90, longitude: 12.49, fillKey: 'city', radius: 5 },
-      { name: 'Miami', latitude: 25.76, longitude: -80.19, fillKey: 'city', radius: 5 },
-      { name: 'Pawnee', latitude: 36.33, longitude: -96.80, fillKey: 'city', radius: 5 }
-    ])
   }
 })
 export default class extends Vue { }
@@ -239,41 +200,6 @@ main {
     font-weight: 200;
     letter-spacing: 0.3em;
   }
-
-  .page-section {
-    width: calc(100% - 4rem);
-    height: auto;
-    margin-left: auto;
-    margin-right: auto;
-  }
-
-  .copy-container {
-    position: relative;
-  }
-
-  .map-tip {
-    position: relative;
-    font-size: 0.8rem;
-    font-weight: 200;
-    color: white;
-    background-color: var(--secondary-color);
-    padding: 0.2rem 0.6rem;
-    transform: translateX(-50%);
-
-    &::before {
-      content: "";
-      display: inline-block;
-      position: absolute;
-      bottom: 100%;
-      left: 50%;
-      width: 0;
-      height: 0;
-      font-size: 0;
-      border: 8px solid transparent;
-      border-bottom: 8px solid var(--secondary-color);
-      transform: translate(-50%);
-    }
-  }
 }
 
 #meet-the-advocates {
@@ -297,10 +223,6 @@ main {
     .advocate-card {
       width: 100%;
       border: 1px solid var(--secondary-color);
-    }
-
-    & > *:nth-child(3n - 1) {
-
     }
   }
 }
